+++ ansibleroletest/cli/test.py
@@ -1,18 +1,9 @@
-from __future__ import print_function
-
 import click
-import logging
-import os
-import six
 import sys
-import yaml
 
-from .container import ContainerManager
-from .docker import client as docker_client
-from .framework import TestFramework
-
-logging.captureWarnings(True)
-
+from ansibleroletest.container import ContainerManager
+from ansibleroletest.docker import client as docker_client
+from ansibleroletest.framework import TestFramework
 
 @click.command(context_settings={'help_option_names': ['-h', '--help']})
 # path options
@@ -65,10 +56,11 @@ logging.captureWarnings(True)
               help='Run test containers in privileged mode (dangerous)')
 @click.option('--cache', is_flag=True,
               help='Cache yum/apt folders on the host')
-@click.option('--save-failed/--no-save-failed', default=False,
-              help='Save failed containers for inspection')
+@click.option('--save', default=None, type=click.Choice(['failed', 'successful', 'all']),
+              help='Save containers, can be either one of "failed", '
+                   '"successful" and "all"')
 @click.argument('role')
-def main(role,
+def test(role,
          config,
          # path args
          roles_path, library_path, plugins_action_path,
@@ -76,9 +68,9 @@ def main(role,
          # ansible-playbook args
          extra_vars, limit, skip_tags, tags, verbosity,
          # misc
-         ansible_version, privileged, cache, save_failed):
+         ansible_version, privileged, cache, save):
     """
-    ansible-role-test is a docker based testing utility for ansible roles.
+    Run tests
 
     ROLE can be either be a local path, a git repository or an ansible-galaxy
     role name.
@@ -106,16 +98,15 @@ def main(role,
             verbosity=verbosity,
             privileged=privileged,
             cache=cache,
-            save_failed=save_failed
+            save=save
         )
 
-        if res != 0 and not save_failed:
+        if res != 0 and save != 'failed':
             click.secho('''
 info: some of the tests have failed. If you wish to inspect the failed
-      containers, rerun the command while adding the --save-failed flag
+      containers, rerun the command while adding the --save=failed flag
       to your command line.''', fg='blue')
     sys.exit(res)
-
 
 def _load_config(conf, config_file=None):
     if not config_file:
@@ -133,6 +124,3 @@ def _load_config(conf, config_file=None):
                 obj_to[k] = os.path.join(base, obj_from[k])
 
     _update(content, conf)
-
-if __name__ == '__main__':
-    main()